Remove unused JSON files and refactor logic evaluator and answer transformer functions
