--- answers.json
@@ -1,6 +1,0 @@
-{
-    "answers": [
-        "The policy can be renewed within a Grace Period of 30 days to maintain continuity of benefits without a break in policy. However, coverage is not available during this grace period.",
-        "The waiting period for cataract surgery is two years."
-    ]
-}
