--- qa_results_optimized.json
@@ -1,10 +1,0 @@
-[
-    {
-        "question": "What is the grace period for premium payment under the National Parivar Mediclaim Plus Policy?",
-        "answer": "{\n  \"answer\": \"The policy can be renewed within a Grace Period of 30 days to maintain continuity of benefits without a break in policy. However, coverage is not available during this grace period.\"\n}"
-    },
-    {
-        "question": "What is the waiting period for cataract surgery?",
-        "answer": "{\"answer\": \"The waiting period for cataract surgery is two years.\"}"
-    }
-]
